wof: move overlay link to game settings
Fixes #1507
