+++ public/pages/games.html
@@ -20,7 +20,16 @@
           <h6>{{ category }}</h6>
           <div class="card mb-2" >
             <div class="card-body">
-              <div v-if="typeof value === 'object'" class="p-0 pl-2 pr-2 " v-for="(currentValue, defaultValue) of value">
+              <div v-if="typeof value === 'object' && (typeof value[defaultValue] === 'undefined' || value[defaultValue] === null)" class="p-0 pl-2 pr-2 " v-for="(currentValue, defaultValue) of ui[category]">
+                <template v-if="ui[category][defaultValue].type === 'link'">
+                  <a :href="currentValue.href" class="mt-1 mb-1" :class="currentValue.class" :target="currentValue.target">
+                    <template v-if="currentValue.rawText">{{ currentValue.rawText }}</template>
+                    <template v-else>{{ commons.translate(currentValue.text) }}</template>
+                  </a>
+                </template>
+              </div>
+              <div v-if="typeof value === 'object' && !defaultValue.startsWith('_') && currentValue !== null" class="p-0 pl-2 pr-2 " v-for="(currentValue, defaultValue) of value">
+
                 <template v-if="selectedGames === 'wheeloffortune' && category === 'options'">
                   <wheel-of-fortune-responses
                     :options="JSON.parse(currentValue)"
@@ -141,6 +150,7 @@
         listOfGames: [],
         selectedGames: null,
         settings: { settings: {} },
+        ui: {},
         isDataChanged: false,
 
         state: {
@@ -171,10 +181,12 @@
           this.settings = { settings: {} }
           this.state.settings = 0
           this.state.loading = 1
-          this.socket[system].emit('settings', (err, settings) => {
+          this.socket[system].emit('settings', (err, settings, ui) => {
             if (system !== this.selectedGames) return
             this.state.loading = 0
             let _settings = _(settings).toPairs().value()
+
+            this.ui = ui
 
             // sorting
             // enabled is first
@@ -192,6 +204,15 @@
             for (let [name, value] of _(_settings.filter(o => o[0] === 'commands')).value()) {
               this.settings[name] = value
             }
+
+            // check if all ui is in settings to be shown
+            for (let key of Object.keys(this.ui)) {
+              if (!this.settings[key]) this.settings[key] = {}
+              for (let k of Object.keys(this.ui[key])) {
+                this.settings[key][k] = this.settings[key][k] || null
+              }
+            }
+
             this.isDataChanged = false
           })
         },
